refactor: commentService 분리
